backend/requirements.txt
fastapi
uvicorn
gunicorn
pydantic
python-dotenv